% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{neuronbridge_avoid}
\alias{neuronbridge_avoid}
\title{Get LM-EM matches for one set of neurons while avoiding another set}
\usage{
neuronbridge_avoid(search, avoid, version = "v2_1_1", threshold = 23000)
}
\arguments{
\item{search}{data item IDs (in \code{neuronbridge_ids}) for which you want to find hits.}

\item{avoid}{data item IDs (in \code{neuronbridge_ids}) that you would rather not have in your hits.}

\item{version}{the precomputed scores to search. For example, \code{"v2_1_1"} refers to this \href{https://neuronbridge.janelia.org/releasenotes/DATA}{release}.}

\item{threshold}{LM-EM matches with a \code{normalizedScore} below this value are not returned.}
}
\value{
a \code{data.frame} of hits. Each row idnciates a separate MIP file with its own \code{nb.id}. The \code{data.frame} is already ranked by \code{normalizedScore}.
Top scores (better match) are at the top of the data frame. The columns mean:
\itemize{

  \item{"publishedName"} { - the \code{id} for the potential hit neuron/line. I.e. specififes a genetic driver resource or a connectome neuron.
  these are the same ids that can be seen with \code{\link{neuronbridge_ids}}.}

  \item{"libraryName"}{ - the data set from which this data item came.}

  \item{"imageURL"}{ - the path on https://s3.amazonaws.com/, at which one can find a 'high res' .png of the MIP file for this data item.}

  \item{"thumbnailURL"}{ - the path on https://s3.amazonaws.com/, at which one can find a 'low res' .jpg thumbnail of the MIP file for this data item.}

  \item{"slideCode"}{ - the unique identifier for the sample from which the MIP came. The first number indicates the date the image was taken by FlyLight.}

  \item{"objective"}{ - the magnification under which the imasge was taken.}

  \item{"gender"}{ - the sex of the fly brain which this data item derives. f = female, m = male.}

  \item{"anatomicalArea"}{ - the gross part of the nervous system images, e.g. brain or ventral nervous system.}

  \item{"alignmentSpace"}{ - the template brain to which the image that formed this MIP, was aligned.
  Typically, this is the JR2018 standard template brain from \href{https://doi.org/10.1101/376384}{Bogovic et al. 2018}.}

  \item{"channel"}{ - number of the channel from the aligned image stack that is represented by this MIP.}

  \item{"mountingProtocol"}{ - the protocol used to prepare brain sample for imaging.}

  \item{"matchingPixels"}{ - the number of overlapping pixels between query (\code{searched.id}) and taregt (\code{nb.id}).}

  \item{"gradientAreaGap "}{ - unsure, seeking clarification from NeuronBridge}

  \item{"normalizedGapScore"}{ - unsure, seeking clarification from NeuronBridge}

  \item{"normalizedScore"}{ - the matching score, created by examining the overlapped pixel number and color depth.
  If the color and xy position of the pixel match between the mask and the searching data,
  then the approach here will count it as a positive matching score}

  \item{"searched.id"}{ - the \code{nb.id} you searched with, i.e. given to the function call}

  \item{"nb.id"}{ - the 'NeuronBridge ID' for the MIP file.}

}
}
\description{
This function uses \code{\link{neuronbridge_search}} to get potential matches for data items (i.e. neurons/lines) that you wish to search
(\code{search}). It also fetches the same information on data items you would rather not have in your 'hits', i.e.
hemibrain connectome neurons that you do not want to 'also'-be in the lines that come up when you search for you neurons in \code{search}. Depending
on your use-case, you may want to read a series of table with \code{\link{neuronbridge_search}} rather than use this function, to filter your matches.
See details below.
}
\examples{
\donttest{
\dontrun{

# Get helpful package with some classe hemibrain neuron IDs
if (!require("hemibrainr")) remotes::install_github("flyconnectome/hemibrainr")

# Load package
library(hemibrainr)

# So this is the 'olfactory PN' neuron we want
search = "542634818"

# And we do not want these other PNs
avoid = setdiff(hemibrainr::pn.ids,search)

# Let's see what we get for ot
hits = neuronbridge_avoid(search = search, avoid = avoid)
# View(hits[1:20,])

}}
}
\seealso{
\code{\link{neuronbridge_info}},
  \code{\link{neuronbridge_mip}},
  \code{\link{neuronbridge_search}}
}
