% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{neuronbridge_hits}
\alias{neuronbridge_hits}
\alias{neuronbridge_search}
\title{Get EM-LM neuron matches for fly brain neurons from neuronbridge.janelia.org}
\usage{
neuronbridge_hits(nb.id, version = "v2_1_1")

neuronbridge_search(
  id,
  version = "v2_1_1",
  dataset = c("detect", "by_line", "by_body"),
  threshold = 10000
)
}
\arguments{
\item{nb.id}{an internal ID used by \href{https://neuronbridge.janelia.org/}{neuronbridge.janelia.org} to identify neurons/lines. This
differs from \code{id} used by other functions, and can be found using \code{\link{neuronbridge_info}}.}

\item{version}{the precomputed scores to search. For example, \code{"v2_1_1"} refers to this \href{https://neuronbridge.janelia.org/releasenotes/DATA}{release}.}

\item{id}{character vector. An identifier/identifiers for the neuron(s)/line(s) you wish to search. This can either be a line code
for a GAL4 line (e.g. \href{https://v2.virtualflybrain.org/org.geppetto.frontend/geppetto?id=VFBexp_FBtp0057908&i=VFB_00017894}{R16F12})
or a split GAL4 line (e.g. \href{https://splitgal4.janelia.org/cgi-bin/view_splitgal4_imagery.cgi?line=LH173}{LH173})
or a 'body ID' for a neuron from the \href{https://neuprint.janelia.org/help/videos?dataset=hemibrain}{hemibrain connectome},
(e.g. \href{https://neuprint.janelia.org/view?bodyid=1537331894}{1537331894}).}

\item{dataset}{whether the ID given is a body ID (\code{"by_body"}) from the hemibrain connectome
or a genetic driver line (\code{"by_line"}). If left at \code{"detect"} then \code{neuronbridger} tries to guess to which \code{id} belongs by using
\code{neuronbridger:::guess_dataset}.}

\item{threshold}{LM-EM matches with a \code{normalizedScore} below this value are not returned. If set to \code{NULL}, the results are not filtered.}
}
\value{
a \code{data.frame} of hits. Each row idnciates a separate MIP file with its own \code{nb.id}. The \code{data.frame} is already ranked by \code{normalizedScore}.
Top scores (better match) are at the top of the data frame. The columns mean:
\itemize{

  \item{"publishedName"} { - the \code{id} for the potential hit neuron/line. I.e. specififes a genetic driver resource or a connectome neuron.
  these are the same ids that can be seen with \code{\link{neuronbridge_ids}}.}

  \item{"libraryName"}{ - the data set from which this data item came.}

  \item{"imageURL"}{ - the path on https://s3.amazonaws.com/, at which one can find a 'high res' .png of the MIP file for this data item.}

  \item{"thumbnailURL"}{ - the path on https://s3.amazonaws.com/, at which one can find a 'low res' .jpg thumbnail of the MIP file for this data item.}

  \item{"slideCode"}{ - the unique identifier for the sample from which the MIP came. The first number indicates the date the image was taken by FlyLight.}

  \item{"objective"}{ - the magnification under which the imasge was taken.}

  \item{"gender"}{ - the sex of the fly brain which this data item derives. f = female, m = male.}

  \item{"anatomicalArea"}{ - the gross part of the nervous system images, e.g. brain or ventral nervous system.}

  \item{"alignmentSpace"}{ - the template brain to which the image that formed this MIP, was aligned.
  Typically, this is the JR2018 standard template brain from \href{https://doi.org/10.1101/376384}{Bogovic et al. 2018}.}

  \item{"channel"}{ - number of the channel from the aligned image stack that is represented by this MIP.}

  \item{"mountingProtocol"}{ - the protocol used to prepare brain sample for imaging.}

  \item{"matchingPixels"}{ - the number of overlapping pixels between query (\code{searched.id}) and taregt (\code{nb.id}).}

  \item{"gradientAreaGap "}{ - unsure, seeking clarification from NeuronBridge}

  \item{"normalizedGapScore"}{ - unsure, seeking clarification from NeuronBridge}

  \item{"normalizedScore"}{ - the matching score, created by examining the overlapped pixel number and color depth.
  If the color and xy position of the pixel match between the mask and the searching data,
  then the approach here will count it as a positive matching score}

  \item{"searched.id"}{ - the \code{nb.id} you searched with, i.e. given to the function call}

  \item{"nb.id"}{ - the 'NeuronBridge ID' for the MIP file.}

}
}
\description{
The function \code{neuronbridge_hits} retrieves a ranked list of hits for a MIP file search.
Specify the MIP file you want to search for by using its \code{nb.id}. This can be found by
using \code{\link{neuronbridge_info}}. \code{neuronbridge_search} will run \code{neuronbridge_info} and then \code{neuronbridge_hits}, to make things easier.
The function \code{neuronbridge_search} can also handle being given multiple IDs, i.e. searching for multiple data items at once.
MIP options can be fetched and visualised with \code{\link{neuronbridge_mip}}.
}
\examples{
\donttest{
\dontrun{
nb.info.em = neuronbridge_info("542634818")
nb.hits = neuronbridge_hits(nb.info.em$nb.id[1])
# View(nb.hits[1:10,]) # see the top 10 hits

# One can also do this in one with:
nb.search = neuronbridge_search("542634818")
## However, this differs from the above in that every MIP
## associated with the given id, is searched. For a connectome
## neuron this is just 1, but for a GAL4 line with MCFO data it
## can be many quite different images.

# Note that nb.info.em here is actually an attribute you can see
nb.info.em = attr(nb.search,"search")
}}
}
\seealso{
\code{\link{neuronbridge_info}},
  \code{\link{neuronbridge_mip}},
  \code{\link{neuronbridge_ids}}
}
