% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{neuronbridge_ids}
\alias{neuronbridge_ids}
\title{Get identifiers for neurons/lines with NeuronBridge results}
\usage{
neuronbridge_ids(version = "v2_1_1")
}
\arguments{
\item{version}{the precomputed scores to search. For example, \code{"v2_1_1"} refers to this \href{https://neuronbridge.janelia.org/releasenotes/DATA}{release}.}
}
\value{
a \code{vector} of identifiers
}
\description{
This function gets all of the neuron identifiers or line codes (e.g.R24A02 is a generation 1 GMR GLA4 line, LH173 is a split-GAL4 line, 5813055865 is a hemibrain neuron)
for published neurons/lines for which NeuronBridge has precomputed matching results. If you query is not among these identifiers you will not be able to get results for it.
}
\examples{
\donttest{
all.nb.ids = neuronbridge_ids()
length(all.nb.ids)
}
}
\seealso{
\code{\link{neuronbridge_hits}},
  \code{\link{neuronbridge_info}},
  \code{\link{neuronbridge_search}}
}
