% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{neuronbridge_info}
\alias{neuronbridge_info}
\title{Get information on neurons and lines used by NeuronBridge}
\usage{
neuronbridge_info(
  id,
  dataset = c("detect", "by_line", "by_body"),
  version = "v2_1_1",
  ...
)
}
\arguments{
\item{id}{character vector. An identifier/identifiers for the neuron(s)/line(s) you wish to search. This can either be a line code
for a GAL4 line (e.g. \href{https://v2.virtualflybrain.org/org.geppetto.frontend/geppetto?id=VFBexp_FBtp0057908&i=VFB_00017894}{R16F12})
or a split GAL4 line (e.g. \href{https://splitgal4.janelia.org/cgi-bin/view_splitgal4_imagery.cgi?line=LH173}{LH173})
or a 'body ID' for a neuron from the \href{https://neuprint.janelia.org/help/videos?dataset=hemibrain}{hemibrain connectome},
(e.g. \href{https://neuprint.janelia.org/view?bodyid=1537331894}{1537331894}).}

\item{dataset}{whether the ID given is a body ID (\code{"by_body"}) from the hemibrain connectome
or a genetic driver line (\code{"by_line"}). If left at \code{"detect"} then \code{neuronbridger} tries to guess to which \code{id} belongs by using
\code{neuronbridger:::guess_dataset}.}

\item{version}{the precomputed scores to search. For example, \code{"v2_1_1"} refers to this \href{https://neuronbridge.janelia.org/releasenotes/DATA}{release}.}

\item{...}{methods passed to \code{neuronbridger:::neuronbridge_fetch}.}
}
\value{
a \code{data.frame} where each row corresponds to a MIP file that has been used to compute matching scores.
Each relates to the \code{id} you gave in the function call, but each also has its own unique \code{nb.id}.
\itemize{

  \item{"publishedName"} { - the \code{id} you gave in the function call.}

  \item{"libraryName"}{ - the data set from which this data item came.}

  \item{"imageURL"}{ - the path on https://s3.amazonaws.com/, at which one can find a 'high res' .png of the MIP file for this data item.}

  \item{"thumbnailURL"}{ - the path on https://s3.amazonaws.com/, at which one can find a 'low res' .jpg thumbnail of the MIP file for this data item.}

  \item{"gender"}{ - the sex of the fly brain which this data item derives}

  \item{"nb.id"}{ - the 'NeuronBridge ID' for the MIP file.}

  \item{"no.images"}{ - the total number of MIP files (each with its own nb.id) that related to the \code{id} for which you searched.}

}
}
\description{
This function gets information stored on \href{https://neuronbridge.janelia.org/}{NeuronBridge}
for *D. melanogaster* neurons and genetic driver lines that have been used in EM-LM searches.
}
\examples{
\donttest{
\dontrun{
# Get information on a mushroombody related GMR GAL4 driver line
nb.info.lm = neuronbridge_info("MB543B")

# Get information on the hemibrain neuron,
nb.info.em = neuronbridge_info("542634818")
}}
}
\seealso{
\code{\link{neuronbridge_hits}},
  \code{\link{neuronbridge_line_contents}},
  \code{\link{neuronbridge_mip}}
}
