% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire.R
\name{ngl_scene_to_nrrd}
\alias{ngl_scene_to_nrrd}
\alias{root_id_to_nrrd}
\title{Make a .nrrd file from neurons in a given flywire scene}
\usage{
ngl_scene_to_nrrd(
  scene,
  reference = "JRC2018U_HR",
  savefolder = file.path(getwd(), "nrrd"),
  file = NULL
)

root_id_to_nrrd(
  root_ids,
  reference = "JRC2018U_HR",
  savefolder = file.path(getwd(), "nrrd"),
  file = NULL
)
}
\arguments{
\item{scene}{a neuroglancer link to a flywire scene. You can get this by hitting the 'share' button in the flywire web browser GUI.}

\item{reference}{the templatebrain to transform the flywire neuron(s) into before creating the .nrrd file. Defaults to \code{"JRC2018U_HR"}, which is
the brain space into which a lot of data for neuronbridge is formatted.}

\item{savefolder}{a folder in which to save .nrrd files. defaults to a folder called 'nrrd' that it will create in your working directory.}

\item{file}{the name of your .nrrd file. Defaults to the root IDs for the given flywire neurons, concatenated with the name of the reference brain space.}

\item{root_ids}{a vector of root IDs that specify flywire meshes.}
}
\value{
a .nrrd file.
}
\description{
Creates a common image file (a .nrrd file) readable in Fiji from a \href{https://flywire.ai/}{flywire} scene. User must have access to the
flywire.ai production data set. See R package \href{https://natverse.org/fafbseg/}{fafbseg} scene for more information on accessing flywire data from R.
}
\examples{
\donttest{
# root_id_to_nrrd(root_ids = c("720575940632668823", "720575940644278051"))
# ascending neurons,
# if this link does not work for you,
# you do not yet have access to the data set
}
}
\seealso{
\code{\link{nrrd_to_mip}}
}
