% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire.R
\name{nrrd_to_mip}
\alias{nrrd_to_mip}
\title{Open Fiji on your computer and use it to convert your .nrrd files into color MIPs}
\usage{
nrrd_to_mip(fiji.path = NULL, MinMem = MaxMem, MaxMem = "2500m")
}
\arguments{
\item{fiji.path}{path to your system's Fiji. If NULL, attempts to find using \code{getOption("jimpipeline.fiji")}.}

\item{MinMem}{lower memory limit}

\item{MaxMem}{upper memory limit}
}
\value{
a MIP files, saved as .tif, converted from a folder of selected .nrrd files
}
\description{
Open Fiji on your computer and use it to convert your .nrrd files into color MIPs for searches. You must have downloaded Fiji onto your machine and also have followed the instructions
\href{https://github.com/JaneliaSciComp/ColorMIP_Mask_Search}{here (search tool)} and \href{https://github.com/JaneliaSciComp/ColorMIP_Mask_Search/tree/master/ColorDepthMIP_Generator}{here (MIP generator)} to install the Fiji plugins
from Scientific Computing at Janelia Research Campus, for the Color MIP Search. When running the MIP generator, select directory that contains 3D stacks to be color depth MIP
then select directory for saving MIPs. Pay attention to the parameters you can then adjust. Skeleton bolded MIP can be useful for data coming from skeleton or even mesh sources.
}
\examples{
\donttest{
nrrd_to_mip(fiji.path = "C:/Program Files/Fiji.app/ImageJ-win64.exe")
}
}
\seealso{
\code{\link{nrrd_to_mip}}
}
